+++ src/main/java/com/programmers/smrtstore/domain/product/application/dto/res/ProductThumbnailResponse.java
@@ -14,7 +14,7 @@ public class ProductThumbnailResponse {
     private Integer price;
     private boolean discountYn;
     private Integer salePrice;
-    private Float discountRatio;
+    private Integer discountRatio;
     private URL thumbnail;
